# app/api/entities/register/route.js
import { NextResponse } from 'next/server';
import { getServiceClient, generateApiKey } from '@/lib/supabase';
import { computeReceiptComposite } from '@/lib/scoring';
import { epProblem } from '@/lib/errors';

/**
 * POST /api/entities/register
 *
 * Register a new entity on the EMILIA Protocol.
 * Returns an API key that the entity uses for all future interactions.
 *
 * Body: {
 *   entity_id: "rex-booking-v2",
 *   display_name: "Rex — Inbound AI Receptionist",
 *   entity_type: "agent" | "merchant" | "service_provider",
 *   description: "Handles inbound SMS booking for service businesses",
 *   capabilities: ["inbound_booking", "sms_reply"],
 *   category: "salon",                    // optional, for merchants
 *   pricing_model: "per_task",            // optional
 *   pricing_amount_cents: 25,             // optional
 *   a2a_endpoint: "https://...",          // optional
 *   ucp_profile_url: "https://...",       // optional
 *   input_schema: { ... },               // optional JSON Schema
 *   output_schema: { ... },              // optional JSON Schema
 * }
 *
 * Returns: { entity, api_key }
 *   api_key is returned ONCE at registration. Store it securely.
 */
export async function POST(request) {
  try {
    const body = await request.json();
    const supabase = getServiceClient();

    // Validate required fields
    const required = ['entity_id', 'display_name', 'entity_type', 'description'];
    for (const field of required) {
      if (!body[field]) {
        return epProblem(400, 'missing_field', `Missing required field: ${field}`);
      }
    }

    const VALID_ENTITY_TYPES = [
      // Commerce entities
      'agent', 'merchant', 'service_provider',
      // Software entities (EP-SX)
      'github_app', 'github_action', 'mcp_server', 'npm_package',
      'chrome_extension', 'shopify_app', 'marketplace_plugin', 'agent_tool',
    ];

    if (!VALID_ENTITY_TYPES.includes(body.entity_type)) {
      return epProblem(400, 'invalid_entity_type', `entity_type must be one of: ${VALID_ENTITY_TYPES.join(', ')}`);
    }

    if (typeof body.display_name === 'string' && body.display_name.length > 200) {
      return epProblem(400, 'display_name_too_long', 'display_name must not exceed 200 characters');
    }

    // Check for duplicate entity_id
    const { data: existing } = await supabase
      .from('entities')
      .select('id')
      .eq('entity_id', body.entity_id)
      .single();

    if (existing) {
      return epProblem(409, 'entity_id_taken', `entity_id "${body.entity_id}" is already registered`);
    }

    // Rate limiting handled by middleware on all /api/* routes

    // Generate embedding from description + capabilities
    let embedding = null;
    if (process.env.OPENAI_API_KEY) {
      const embeddingText = [
        body.display_name,
        body.description,
        ...(body.capabilities || []),
        body.category,
      ].filter(Boolean).join('. ');

      const embRes = await fetch('https://api.openai.com/v1/embeddings', {
        method: 'POST',
        headers: {
          'Authorization': `Bearer ${process.env.OPENAI_API_KEY}`,
          'Content-Type': 'application/json',
        },
        body: JSON.stringify({
          model: 'text-embedding-3-small',
          input: embeddingText,
        }),
      });

      if (embRes.ok) {
        const embData = await embRes.json();
        embedding = embData.data[0].embedding;
      }
    }

    // Generate API key
    const { key: apiKey, hash: apiKeyHash, prefix } = generateApiKey();

    // Derive owner_id from client IP hash (not caller-supplied — prevents spoofing)
    const clientIP = request.headers.get('x-forwarded-for')?.split(',')[0]?.trim()
      || request.headers.get('x-real-ip') || 'unknown';
    const ownerHash = Array.from(
      new Uint8Array(await crypto.subtle.digest('SHA-256', new TextEncoder().encode(clientIP)))
    ).map(b => b.toString(16).padStart(2, '0')).join('').slice(0, 32);

    // Insert entity
    const { data: entity, error: insertError } = await supabase
      .from('entities')
      .insert({
        entity_id: body.entity_id,
        owner_id: `ip_${ownerHash}`,
        display_name: body.display_name,
        entity_type: body.entity_type,
        description: body.description,
        website_url: body.website_url || null,
        capabilities: body.capabilities || [],
        input_schema: body.input_schema || null,
        output_schema: body.output_schema || null,
        category: body.category || null,
        software_meta: body.software_meta || null,
        service_area: body.service_area || null,
        pricing_model: body.pricing_model || null,
        pricing_amount_cents: body.pricing_amount_cents || 0,
        capability_embedding: embedding,
        a2a_endpoint: body.a2a_endpoint || null,
        ucp_profile_url: body.ucp_profile_url || null,
        api_key_hash: apiKeyHash,
      })
      .select()
      .single();

    if (insertError) {
      console.error('Entity insert error:', insertError);
      return epProblem(500, 'registration_failed', 'Failed to register entity');
    }

    // Create API key record
    await supabase.from('api_keys').insert({
      entity_id: entity.id,
      key_hash: apiKeyHash,
      key_prefix: prefix,
      label: 'Default key',
    });

    return NextResponse.json({
      entity: {
        id: entity.id,
        entity_id: entity.entity_id,
        display_name: entity.display_name,
        entity_type: entity.entity_type,
        compat_score: entity.emilia_score,
        confidence: 'pending',
        status: entity.status,
        created_at: entity.created_at,
      },
      api_key: apiKey,
      message: 'Store this API key securely. It will not be shown again.',
      _note: 'Query /api/trust/profile/:entityId for full trust profile. compat_score is for sorting only.',
    }, { status: 201 });
  } catch (err) {
    console.error('Registration error:', err);
    return epProblem(500, 'internal_error', 'Internal server error');
  }
}