# openapi.yaml
openapi: 3.0.3
info:
  title: EMILIA Protocol API
  description: |
    Trust profiles, policy evaluation, install preflight, disputes, and appeals
    for counterparties, software, and machine actors.

    EMILIA = Evidence-based Mediation & Integrity Layer for Interactions and Appeals.

    Constitutional principle: trust must never be more powerful than appeal.
  version: 1.1.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    email: team@emiliaprotocol.ai
    url: https://emiliaprotocol.ai
servers:
  - url: https://emiliaprotocol.ai
    description: Production
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/entities/register:
    post:
      tags: [Entities]
      summary: Register a new entity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [entity_id, display_name, entity_type]
              properties:
                entity_id:
                  type: string
                display_name:
                  type: string
                entity_type:
                  type: string
                  enum: [agent, merchant, service_provider, github_app, github_action, mcp_server, npm_package, chrome_extension, shopify_app, marketplace_plugin, agent_tool]
                description:
                  type: string
      responses:
        '201':
          description: Entity registered
  /api/entities/search:
    get:
      tags: [Entities]
      summary: Search entities
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: min_confidence
          in: query
          schema:
            type: string
            enum: [pending, insufficient, provisional, emerging, confident]
      responses:
        '200':
          description: Search results
...
  /api/disputes/report:
    post:
      tags: [Disputes]
      summary: File a human trust report (no auth required)
      description: |
        Human appeal endpoint. No authentication required.
        "EP must never make trust more powerful than appeal."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [entity_id, report_type]
              properties:
                entity_id:
                  type: string
                report_type:
                  type: string
                  enum: [wrongly_downgraded, harmed_by_trusted_entity, fraudulent_entity, inaccurate_profile, fake_receipts, unsafe_software, misleading_identity, terms_violation, demo_challenge, other]
                description:
                  type: string
                contact_email:
                  type: string
      responses:
        '201':
          description: Report received
...
  /api/delegations/create:
    post:
      tags: [Delegations]
      summary: Create a delegation
      description: |
        Creates a principal-agent delegation. The principal authorizes an agent
        entity to act on their behalf within a defined scope, with optional
        value limits and expiry.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [principal_id, agent_entity_id, scope]
              properties:
                principal_id:
                  type: string
                agent_entity_id:
                  type: string
                scope:
                  type: array
                  items:
                    type: string
                  description: List of permitted action types
                max_value_usd:
                  type: number
                  description: Optional value cap in USD
                expires_at:
                  type: string
                  format: date-time
                constraints:
                  type: object
                  description: Additional constraints on the delegation
      responses:
        '201':
          description: Delegation created
        '401':
          description: Authentication required
...
  /api/audit:
    get:
      tags: [System]
      summary: Query the audit trail
      description: |
        Query the append-only audit trail. Requires authentication.
        Every trust-changing action is recorded with before/after state.
      security:
        - BearerAuth: []
      parameters:
        - name: target_id
          in: query
          schema:
            type: string
          description: Filter by target entity/dispute/receipt ID
        - name: target_type
          in: query
          schema:
            type: string
          description: Filter by target type (e.g. "dispute", "receipt", "entity")
        - name: actor_id
          in: query
          schema:
            type: string
          description: Filter by the actor who performed the action
        - name: event_type
          in: query
          schema:
            type: string
          description: Filter by event type
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Audit events
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items:
                      type: object
                  count:
                    type: integer
                  offset:
                    type: integer
                  limit:
                    type: integer
        '401':
          description: Authentication required

  # ===========================================================================
  # Internal / Cron
  # ===========================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: EP API key (ep_live_...)
    CronAuth:
      type: http
      scheme: bearer
      description: CRON_SECRET for internal/operator endpoints
...